+++ rhine-examples/src/Demonstration.hs
@@ -1,36 +1,33 @@
-{-# LANGUAGE DataKinds        #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE RankNTypes       #-}
-{-# LANGUAGE TypeFamilies     #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE TypeFamilies #-}
 
 import FRP.Rhine
-import FRP.Rhine.Clock.Realtime.Millisecond
-import FRP.Rhine.Schedule.Concurrently
-import FRP.Rhine.ResamplingBuffer.Collect
 
--- | Create a simple message containing the time stamp since program start,
+-- | Create a simple message containing the time stamp since initialisation,
 --   for each tick of the clock.
 --   Since 'createMessage' works for arbitrary clocks (and doesn't need further input data),
 --   it is a 'Behaviour'.
---   @td@ is the 'TimeDomain' of any clock used to sample,
+--   @time@ is the 'TimeDomain' of any clock used to sample,
 --   and it needs to be constrained in order for time differences
 --   to have a 'Show' instance.
 createMessage
-  :: (Monad m, Show (Diff td))
+  :: (Monad m, Show (Diff time))
   => String
-  -> Behaviour m td String
+  -> Behaviour m time String
 createMessage str
-  =   timeInfoOf sinceStart >-> arr show
+  =   timeInfoOf sinceInit >-> arr show
   >-> arr (("Clock " ++ str ++ " has ticked at: ") ++)
 
 -- | Specialise 'createMessage' to a specific clock,
 --   ticking twice a second.
-ms500 :: SyncSF IO (Millisecond 500) () String
+ms500 :: ClSF IO (Millisecond 500) () String
 ms500 = createMessage "500 MS"
 
 -- | Specialise 'createMessage' to a different clock,
 --   ticking every 1.2 seconds.
-ms1200 :: SyncSF IO (Millisecond 1200) () String
+ms1200 :: ClSF IO (Millisecond 1200) () String
 ms1200 = createMessage "1200 MS"
 
 
@@ -38,15 +35,15 @@ ms1200 = createMessage "1200 MS"
 --   Let us assume we want to assure that 'printEverySecond'
 --   is only called every second,
 --   then we constrain its type signature with the clock @Millisecond 1000@.
-printEverySecond :: Show a => SyncSF IO (Millisecond 1000) a ()
-printEverySecond = arrMSync print
+printEverySecond :: Show a => ClSF IO (Millisecond 1000) a ()
+printEverySecond = arrMCl print
 
 -- | Create messages every 500 ms and every 1200 ms,
 --   collecting all of them in a list,
 --   which is output at every second.
 main :: IO ()
 main = flow $
-  ms500 @@ waitClock **@ concurrently @** ms1200 @@ waitClock
+  ms500 @@ waitClock ||@ scheduleMillisecond @|| ms1200 @@ waitClock
   >-- collect -@- concurrently -->
   printEverySecond @@ waitClock