+++ rhine-examples/src/HelloWorld.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE DataKinds #-}
 import FRP.Rhine
-import FRP.Rhine.Clock.Realtime.Millisecond
 
 main :: IO ()
-main = flow $ arrMSync_ (putStrLn "Hello World!") @@ (waitClock :: Millisecond 100)
+main = flow $ constMCl (putStrLn "Hello World!") @@ (waitClock :: Millisecond 100)