+++ rhine/src/FRP/Rhine/ClSF.hs
@@ -1,0 +1,18 @@
+{- |
+Clocked signal functions, i.e. monadic stream functions ('MSF's)
+that are aware of time.
+This module reexports core functionality
+(such as time effects and 'Behaviour's),
+exception handling, reader monad handling,
+and a wealth of utilities such as digital signal processing units.
+Documentation can be found in the individual modules.
+-}
+
+module FRP.Rhine.ClSF ( module X ) where
+
+
+-- rhine
+import FRP.Rhine.ClSF.Core   as X
+import FRP.Rhine.ClSF.Except as X
+import FRP.Rhine.ClSF.Reader as X
+import FRP.Rhine.ClSF.Util   as X