+++ rhine/src/FRP/Rhine/ClSF/Reader.hs
@@ -1,7 +1,11 @@
-{-# LANGUAGE RankNTypes    #-}
+{- |
+Create and remove 'ReaderT' layers in 'ClSF's.
+-}
+
+{-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE TupleSections #-}
-{-# LANGUAGE TypeFamilies  #-}
-module FRP.Rhine.SyncSF.Reader where
+{-# LANGUAGE TypeFamilies #-}
+module FRP.Rhine.ClSF.Reader where
 
 -- base
 import Data.Tuple (swap)
@@ -13,7 +17,7 @@ import Control.Monad.Trans.Reader
 import qualified Control.Monad.Trans.MSF.Reader as MSF
 
 -- rhine
-import FRP.Rhine.SyncSF.Core
+import FRP.Rhine.ClSF.Core
 
 
 -- | Commute two 'ReaderT' transformer layers past each other
@@ -26,7 +30,7 @@ commuteReaders a
 --   by passing the original behaviour the extra @r@ input.
 readerS
   :: Monad m
-  => SyncSF m cl (a, r) b -> SyncSF (ReaderT r m) cl a b
+  => ClSF m cl (a, r) b -> ClSF (ReaderT r m) cl a b
 readerS behaviour
   = liftMSFPurer commuteReaders $ MSF.readerS $ arr swap >>> behaviour
 
@@ -34,12 +38,12 @@ readerS behaviour
 --   by making it an explicit input to the behaviour.
 runReaderS
   :: Monad m
-  => SyncSF (ReaderT r m) cl a b -> SyncSF m cl (a, r) b
+  => ClSF (ReaderT r m) cl a b -> ClSF m cl (a, r) b
 runReaderS behaviour
   = arr swap >>> (MSF.runReaderS $ liftMSFPurer commuteReaders behaviour)
 
 -- | Remove a 'ReaderT' layer by passing the readonly environment explicitly.
 runReaderS_
   :: Monad m
-  => SyncSF (ReaderT r m) cl a b -> r -> SyncSF m cl a b
+  => ClSF (ReaderT r m) cl a b -> r -> ClSF m cl a b
 runReaderS_ behaviour r = arr (, r) >>> runReaderS behaviour