+++ rhine/src/FRP/Rhine/Clock/Realtime/Busy.hs
@@ -1,12 +1,14 @@
+{- | A "'Busy'" clock that ticks without waiting. -}
+
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE TypeFamilies          #-}
+{-# LANGUAGE TypeFamilies #-}
 module FRP.Rhine.Clock.Realtime.Busy where
 
 -- base
 import Data.Time.Clock
 
 -- rhine
-import FRP.Rhine
+import FRP.Rhine.Clock
 
 {- |
 A clock that ticks without waiting.
@@ -16,10 +18,10 @@ side effects, time measurement and framework overhead.
 data Busy = Busy
 
 instance Clock IO Busy where
-  type TimeDomainOf Busy = UTCTime
-  type Tag          Busy = ()
+  type Time Busy = UTCTime
+  type Tag  Busy = ()
 
-  startClock _ = do
+  initClock _ = do
     initialTime <- getCurrentTime
     return
       ( arrM_ getCurrentTime