+++ rhine/src/FRP/Rhine/ResamplingBuffer.hs
@@ -1,7 +1,19 @@
-{-# LANGUAGE RankNTypes      #-}
+{- |
+This module introduces 'ResamplingBuffer's,
+which are primitives that consume and produce data at different rates.
+Just as schedules form the boundaries between different clocks,
+(resampling) buffers form the boundaries between
+synchronous signal functions ticking at different speeds.
+-}
+
+{-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE RecordWildCards #-}
-{-# LANGUAGE TypeFamilies    #-}
-module FRP.Rhine.ResamplingBuffer where
+{-# LANGUAGE TypeFamilies #-}
+module FRP.Rhine.ResamplingBuffer
+  ( module FRP.Rhine.ResamplingBuffer
+  , module FRP.Rhine.Clock
+  )
+  where
 
 -- rhine
 import FRP.Rhine.Clock
@@ -11,10 +23,10 @@ import FRP.Rhine.Clock
 import Control.Arrow (second)
 
 -- A quick note on naming conventions, to whoever cares:
--- . Call a single clock cl.
--- . Call several clocks cl1, cl2 etc. in most situations.
--- . Call it cla, clb etc. when they are Leftmost or Rightmost clocks,
--- i.e. associated to particular boundary types a, b etc.,
+-- . Call a single clock @cl@.
+-- . Call several clocks @cl1@, @cl2@ etc. in most situations.
+-- . Call it @cla@, @clb@ etc. when they are 'In' or 'Out' clocks,
+-- i.e. associated to particular boundary types @a@, @b@ etc.,
 
 {- | A stateful buffer from which one may 'get' a value,
 or to which one may 'put' a value,