+++ rhine/src/FRP/Rhine/ResamplingBuffer/LIFO.hs
@@ -1,8 +1,12 @@
+{- |
+Different implementations of LIFO buffers.
+-}
+
 {-# LANGUAGE RecordWildCards #-}
 module FRP.Rhine.ResamplingBuffer.LIFO where
 
 -- base
-import Prelude hiding (length)
+import Prelude hiding (length, take)
 
 -- containers
 import Data.Sequence
@@ -15,14 +19,23 @@ import FRP.Rhine.ResamplingBuffer.Timeless
 
 -- | An unbounded LIFO buffer.
 --   If the buffer is empty, it will return 'Nothing'.
-lifo :: Monad m => ResamplingBuffer m cl1 cl2 a (Maybe a)
-lifo = timelessResamplingBuffer AsyncMealy {..} empty
+lifoUnbounded :: Monad m => ResamplingBuffer m cl1 cl2 a (Maybe a)
+lifoUnbounded = timelessResamplingBuffer AsyncMealy {..} empty
   where
     amPut as a = return $ a <| as
     amGet as   = case viewl as of
       EmptyL   -> return (Nothing, empty)
       a :< as' -> return (Just a , as'  )
 
+-- |  A bounded LIFO buffer that forgets the oldest values when the size is above a given threshold.
+--   If the buffer is empty, it will return 'Nothing'.
+lifoBounded :: Monad m => Int -> ResamplingBuffer m cl1 cl2 a (Maybe a)
+lifoBounded threshold = timelessResamplingBuffer AsyncMealy {..} empty
+  where
+    amPut as a = return $ take threshold $ a <| as
+    amGet as = case viewl as of
+      EmptyL   -> return (Nothing, empty)
+      a :< as' -> return (Just a , as'  )
 
 -- | An unbounded LIFO buffer that also returns its current size.
 lifoWatch :: Monad m => ResamplingBuffer m cl1 cl2 a (Maybe a, Int)