+++ rhine/src/FRP/Rhine/ResamplingBuffer/Timeless.hs
@@ -1,7 +1,12 @@
+{- |
+Resampling buffers from asynchronous Mealy machines.
+These are used in many other modules implementing 'ResamplingBuffer's.
+-}
+
 {-# LANGUAGE RecordWildCards #-}
 module FRP.Rhine.ResamplingBuffer.Timeless where
 
-import FRP.Rhine
+import FRP.Rhine.ResamplingBuffer
 
 -- | An asynchronous, effectful Mealy machine description.
 --   (Input and output do not happen simultaneously.)