+++ rhine/src/FRP/Rhine/ResamplingBuffer/Util.hs
@@ -1,3 +1,7 @@
+{- |
+Several utilities to create 'ResamplingBuffer's.
+-}
+
 {-# LANGUAGE RankNTypes #-}
 module FRP.Rhine.ResamplingBuffer.Util where
 
@@ -5,41 +9,43 @@ module FRP.Rhine.ResamplingBuffer.Util where
 import Control.Monad.Trans.Reader (runReaderT)
 
 -- rhine
-import FRP.Rhine
+import FRP.Rhine.Clock
+import FRP.Rhine.ClSF
+import FRP.Rhine.ResamplingBuffer
 
 -- * Utilities to build 'ResamplingBuffer's from smaller components
 
 infix 2 >>-^
--- | Postcompose a 'ResamplingBuffer' with a matching 'SyncSF'.
+-- | Postcompose a 'ResamplingBuffer' with a matching 'ClSF'.
 (>>-^) :: Monad m
       => ResamplingBuffer m cl1 cl2 a b
-      -> SyncSF           m     cl2   b c
+      -> ClSF             m     cl2   b c
       -> ResamplingBuffer m cl1 cl2 a   c
-resBuf >>-^ syncSF = ResamplingBuffer put_ get_
+resBuf >>-^ clsf = ResamplingBuffer put_ get_
   where
-    put_ theTimeInfo a = (>>-^ syncSF) <$> put resBuf theTimeInfo a
+    put_ theTimeInfo a = (>>-^ clsf) <$> put resBuf theTimeInfo a
     get_ theTimeInfo   = do
       (b, resBuf') <- get resBuf theTimeInfo
-      (c, syncSF') <- unMSF syncSF b `runReaderT` theTimeInfo
-      return (c, resBuf' >>-^ syncSF')
+      (c, clsf')   <- unMSF clsf b `runReaderT` theTimeInfo
+      return (c, resBuf' >>-^ clsf')
 
 
 infix 1 ^->>
--- | Precompose a 'ResamplingBuffer' with a matching 'SyncSF'.
+-- | Precompose a 'ResamplingBuffer' with a matching 'ClSF'.
 (^->>) :: Monad m
-      => SyncSF           m cl1     a b
+      => ClSF             m cl1     a b
       -> ResamplingBuffer m cl1 cl2   b c
       -> ResamplingBuffer m cl1 cl2 a   c
-syncSF ^->> resBuf = ResamplingBuffer put_ get_
+clsf ^->> resBuf = ResamplingBuffer put_ get_
   where
     put_ theTimeInfo a = do
-      (b, syncSF') <- unMSF syncSF a `runReaderT` theTimeInfo
-      resBuf'      <- put resBuf theTimeInfo b
-      return $ syncSF' ^->> resBuf'
-    get_ theTimeInfo   = second (syncSF ^->>) <$> get resBuf theTimeInfo
+      (b, clsf') <- unMSF clsf a `runReaderT` theTimeInfo
+      resBuf'    <- put resBuf theTimeInfo b
+      return $ clsf' ^->> resBuf'
+    get_ theTimeInfo   = second (clsf ^->>) <$> get resBuf theTimeInfo
 
 
-infix 4 *-*
+infixl 4 *-*
 -- | Parallely compose two 'ResamplingBuffer's.
 (*-*) :: Monad m
       => ResamplingBuffer m cl1 cl2  a      b
@@ -56,6 +62,15 @@ resBuf1 *-* resBuf2 = ResamplingBuffer put_ get_
       (d, resBuf2') <- get resBuf2 theTimeInfo
       return ((b, d), resBuf1' *-* resBuf2')
 
+infixl 4 &-&
+-- | Parallely compose two 'ResamplingBuffer's, duplicating the input.
+(&-&) :: Monad m
+      => ResamplingBuffer m cl1 cl2  a  b
+      -> ResamplingBuffer m cl1 cl2  a     c
+      -> ResamplingBuffer m cl1 cl2  a (b, c)
+resBuf1 &-& resBuf2 = arr (\a -> (a, a)) ^->> resBuf1 *-* resBuf2
+
+
 -- | Given a 'ResamplingBuffer' where the output type depends on the input type polymorphically,
 --   we can produce a timestamped version that simply annotates every input value
 --   with the 'TimeInfo' when it arrived.
@@ -63,4 +78,4 @@ timestamped
   :: Monad m
   => (forall b. ResamplingBuffer m cl clf b (f b))
   -> ResamplingBuffer m cl clf a (f (a, TimeInfo cl))
-timestamped resBuf = (syncId &&& timeInfo) ^->> resBuf
+timestamped resBuf = (clId &&& timeInfo) ^->> resBuf